rename module to spin
