+++ util/src/sync/mod.rs
@@ -6,5 +6,4 @@ pub use loom::sync::atomic;
 #[cfg(all(not(test), not(feature = "loom")))]
 pub use core::sync::atomic;
 
-mod mutex;
-pub use self::mutex::*;
+pub mod spin;
